requirements.txt
streamlit>=1.32.0
plotly>=5.20.0
ccxt>=4.2.0
finta>=1.3
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dateutil>=2.8.2
yfinance>=0.2.36
